requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==1.0.0b0
idna==3.10
Jinja2==3.1.4
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
rfc3986==2.0.0
rich==13.9.4
ruff==0.8.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==1.4.53
starlette==0.41.3
typer==0.13.1
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
